% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID115_functions.R
\name{AnnotateIDVCFsWithTransRanges}
\alias{AnnotateIDVCFsWithTransRanges}
\title{Add sequence context and transcript information to an in-memory ID VCF}
\usage{
AnnotateIDVCFsWithTransRanges(
  ID.vcfs,
  ref.genome,
  trans.ranges = NULL,
  vcf.names = NULL
)
}
\arguments{
\item{ID.vcfs}{A list of in-memory ID VCF as a \code{data.frame}.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{trans.ranges}{Optional. If \code{ref.genome} specifies one of the
\code{\link{BSgenome}} object
\enumerate{
\item \code{\link[BSgenome.Hsapiens.1000genomes.hs37d5]{BSgenome.Hsapiens.1000genomes.hs37d5}}
\item \code{\link[BSgenome.Hsapiens.UCSC.hg38]{BSgenome.Hsapiens.UCSC.hg38}}
\item \code{\link[BSgenome.Mmusculus.UCSC.mm10]{BSgenome.Mmusculus.UCSC.mm10}}
}
then the function will infer \code{trans.ranges} automatically. Otherwise,
user will need to provide the necessary \code{trans.ranges}. Please refer to
\code{\link{TranscriptRanges}} for more details.
If \code{is.null(trans.ranges)} do not add transcript range
information.}

\item{vcf.names}{list of names of the vcfs}
}
\value{
A list of in-memory ID VCFs each a \code{data.table}. These have been annotated
with the sequence context (column name \code{seq.context}) and with
transcript information in the form of a gene symbol (e.g. \code{"TP53"})
and transcript strand. This information is in the columns
\code{trans.start.pos}, \code{trans.end.pos} , \code{trans.strand},
\code{trans.Ensembl.gene.ID} and \code{trans.gene.symbol} in the output.
These columns are not added if \code{is.null(trans.ranges)}.
}
\description{
Add sequence context and transcript information to an in-memory ID VCF
}
\examples{
file <- c(system.file("extdata/Strelka-ID-vcf",
                      "Strelka.ID.GRCh37.s1.vcf",
                      package = "ICAMSxtra"))
list.of.vcfs <- ICAMS::ReadStrelkaIDVCFs(file)
if (requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5", quietly = TRUE)) {
  annotated.ID.vcfs <- AnnotateIDVCFsWithTransRanges(list.of.vcfs, ref.genome = "hg19",
                                                    trans.ranges = ICAMS::trans.ranges.GRCh37)
  }
}
