% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID115_fns.R
\name{Canonicalize1INS115}
\alias{Canonicalize1INS115}
\title{Given an insertion and its sequence context, categorize it.}
\usage{
Canonicalize1INS115(context, ins.sequence, pos, trace = 0, strand)
}
\arguments{
\item{context}{}

\item{ins.sequence}{}

\item{pos}{The position of \code{ins.sequence} in \code{context}.}

\item{trace}{}

\item{strand}{}
}
\value{
A string that is the canonical representation of
the given insertion type.
}
\description{
Given an insertion and its sequence context, categorize it.
}
\details{
same as Canonicalize1INS except when
\code{insertion.size == 1} & \code{rep.count == 0}
in which the format will be \code{INS:Y:1:0_PF}
where Y is C or T, P is the preceding nucleotide, and
F is the following nucleotide
}
\keyword{internal}
