% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID115_functions.R
\name{CanonicalizeID115}
\alias{CanonicalizeID115}
\title{Determine the mutation types of insertions and deletions.}
\usage{
CanonicalizeID115(context, ref, alt, pos, strand)
}
\arguments{
\item{context}{A vector of ample surrounding
sequence on each side the variants}

\item{ref}{Vector of reference alleles}

\item{alt}{Vector of alternative alleles}

\item{pos}{Vector of the positions of the insertions and deletions in
\code{context}.}

\item{strand}{NULL by default. But when called by \code{ID115_StrandBiasGetPlottables},
\code{strand} is either \code{+} or \code{-}. The return value will
include :trans or :nontrans indicating whether the deletion occured on
the transcribed or non-transcribed strand.}
}
\value{
A vector of strings that are the canonical representations
of the given insertions and deletions.
}
\description{
Determine the mutation types of insertions and deletions.
}
\keyword{internal}
