% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID115_fns.R
\name{Collapse115CatalogTo83}
\alias{Collapse115CatalogTo83}
\title{"Collapse" a catalog}
\usage{
Collapse115CatalogTo83(catalog)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}}.}
}
\value{
A catalog as defined in \code{\link{ICAMS}}.
}
\description{
\code{Collapse115CatalogTo83} Collapse a ID 115 catalog
to a ID 83 catalog.
}
