% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID115_functions.R
\name{CollapseID115CatalogsToID83s}
\alias{CollapseID115CatalogsToID83s}
\title{"Collapse" a matrix of ID 115 catalogs to ID 83 catalog}
\usage{
CollapseID115CatalogsToID83s(catalogs)
}
\arguments{
\item{catalogs}{A catalog as defined in \code{\link{ICAMS}}.}
}
\value{
A catalog as defined in \code{\link{ICAMS}}.
}
\description{
"Collapse" a matrix of ID 115 catalogs to ID 83 catalog
}
