% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCh38.proportions}
\alias{GRCh38.proportions}
\title{Genic/intergenic size and proportions for H.sapiens BSgenome GRCh38}
\format{
A list of 5 numbers with the names:
\enumerate{
\item total.bp 2. coding.bp 3. noncoding.bp 4. prop.coding and 5. prop.noncoding
}
}
\usage{
GRCh38.proportions
}
\description{
This data is designed to be used in function
\code{\link{PlotTranscriptionAssociatedDamageToPdf}}
}
\keyword{datasets}
