% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.sigs.R
\name{MatchSigs2Directions}
\alias{MatchSigs2Directions}
\title{Bidirectional closest similarities between two sets of signatures.}
\usage{
MatchSigs2Directions(sigs1, sigs2)
}
\arguments{
\item{sigs1}{Matrix of signatures; colnames are used as signature
identifiers, and the colnames in \code{sigs1} should be distinguishable from
those in \code{sigs2}.}

\item{sigs2}{Matrix of signatures; colnames are used as signature identifiers.}
}
\value{
A list with the elements:

\code{averCosSim}: the average of the cosine similarities between each signature in
\code{sigs1} and its closest match in \code{sigs2} and the closest match
between each signature in \code{sigs2} and its closest match in \code{sigs1}.

\code{match1}: a data frame with rownames being signature identifiers from
\code{sigs1}, the signature identifier of the closest match in \code{sigs1}
in the 1st column, and the cosine similarity between them in the 2nd column.

\code{match2}: a data frame with the row names being signature identifiers
from \code{sigs2}, the signature identifier of the closest match in
\code{sigs1} in the 1st column, and the cosine similarity between them in the
2nd column.

\code{match1} and \code{match2} might not have the same number of rows.
}
\description{
Bidirectional closest similarities between two sets of signatures.
}
\examples{
seta <- matrix(c(1, 3,   4, 1, 2, 4), ncol = 2)
setb <- matrix(c(1, 3.1, 4, 5, 1, 1, 1, 2.8, 4), ncol = 3)
colnames(seta) <- c("A.1", "A.2")
colnames(setb) <- c("B.1", "B.2", "B.3")
MatchSigs2Directions(seta, setb)
}
