% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_error_bar_fns.R
\name{MeanOfSpectraAsSig}
\alias{MeanOfSpectraAsSig}
\title{Return the mean of multiple spectra as a signature}
\usage{
MeanOfSpectraAsSig(
  spectra,
  mean.weighted = TRUE,
  title = "sig.from.spectra.mean"
)
}
\arguments{
\item{spectra}{An \code{\link[ICAMS]{ICAMS}} spectrum catalog.
Convert each spectrum to a signature and then compute the
mean.}

\item{mean.weighted}{Logical. Whether to weigh the samples according to the
number of mutations in them to calculate the weighted mean as the consensus
signature. Default is TRUE. If FALSE, then arithmetic mean will be
calculated.}

\item{title}{The name of the output signature.}
}
\description{
Return the mean of multiple spectra as a signature
}
