% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID115_functions.R
\name{PlotID115AsID83ToPdf}
\alias{PlotID115AsID83ToPdf}
\title{Plot an ID 115 signatures (default) or catalog as standard ID83 and save as pdf file}
\usage{
PlotID115AsID83ToPdf(catalog, file, ylim = NULL)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}}.}

\item{file}{The name of the PDF file to be produced.}

\item{ylim}{Has the usual meaning. Only implemented for SBS96Catalog and
IndelCatalog.}
}
\value{
A list whose first element is a logic value indicating whether the
plot is successful. For \strong{SBS192Catalog} with "counts" catalog.type
and non-null abundance and \code{plot.SBS12 = TRUE}, the list will have a
second element which is a list containing the strand bias statistics.
}
\description{
Plot an ID 115 signatures (default) or catalog as standard ID83 and save as pdf file
}
