% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID115_functions.R
\name{PlotID115Catalog}
\alias{PlotID115Catalog}
\title{Plot \strong{one} spectrum or signature}
\usage{
PlotID115Catalog(catalog, ylim = NULL)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{ylim}{Has the usual meaning. Only implemented for SBS96Catalog and
IndelCatalog.}
}
\description{
Plot the spectrum of \strong{one} sample or plot \strong{one} signature. The
type of graph is based on one attribute("catalog.type") of the input catalog.
You can first use \code{\link{TransformCatalog}} to get different types of
catalog and then do the plotting.
}
