% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID115_fns.R
\name{PlotID115CatalogToPdf}
\alias{PlotID115CatalogToPdf}
\title{Plot catalog to a PDF file}
\usage{
PlotID115CatalogToPdf(catalog, file, ylim = NULL)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{file}{The name of the PDF file to be produced.}

\item{ylim}{Has the usual meaning. Only implemented for SBS96Catalog and IndelCatalog.}
}
\value{
A list whose first element is a logic value indicating whether the
plot is successful. For \strong{SBS192Catalog} with "counts" catalog.type
and non-null abundance and \code{plot.SBS12 = TRUE}, the list will have a
second element which is a list containing the strand bias statistics.
}
\description{
Plot catalog to a PDF file. The type of graph is based on one
attribute("catalog.type") of the input catalog. You can first use
\code{\link{TransformCatalog}} to get different types of catalog and then do
the plotting.
}
\note{
The sizes of repeats involved in deletions range from 0 to 5+ in the
mutational-spectra and signature catalog rownames, but for plotting and
end-user documentation deletion repeat sizes range from 1 to 6+.
}
