% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID_fns.R
\name{PlotTranscriptionAssociatedDamageToPdf}
\alias{PlotTranscriptionAssociatedDamageToPdf}
\title{Plot indel counts on transcribed and nontranscribed strands to pdf}
\usage{
PlotTranscriptionAssociatedDamageToPdf(
  list.of.vcfs,
  ref.genome,
  names.of.vcfs,
  proportions,
  file
)
}
\arguments{
\item{list.of.vcfs}{List of in-memory ID VCFs. The list names will be
the sample ids in the output catalog.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{names.of.vcfs}{list of names of vcfs}

\item{proportions}{The gene proportions for the genome e.g. \code{GRCh37.proportions} or \code{GRCh38.proportions}}

\item{file}{The name of the PDF file to be produced.}
}
\value{
a list of tables of p-values for each vcf
}
\description{
Plot indel counts on transcribed and nontranscribed strands to pdf
}
\section{Note}{

The strand bias statistics are Benjamini-Hochberg q-values based on two-sided
binomial tests of the mutation counts on the transcribed and untranscribed strands
relative to the actual abundances of C and T on the transcribed strand. On the
plot, asterisks indicate q-values as follows *, \eqn{Q<0.05}; **, \eqn{Q<0.01}; ***,
\eqn{Q<0.001}.
}

\examples{
dirs <- c(system.file("extdata/Strelka-ID-vcf", "Strelka.ID.GRCh37.s1.vcf", package="ICAMSxtra"),
          system.file("extdata/Strelka-ID-vcf", "Strelka.ID.GRCh37.s2.vcf", package="ICAMSxtra"))

list.of.vcfs <- ICAMS::ReadStrelkaIDVCFs(dirs, names.of.VCFs = c("s1","s2"))
PlotTranscriptionAssociatedDamageToPdf(list.of.vcfs = list.of.vcfs, 
                                       ref.genome = "hg19", 
                                       names.of.vcfs = c("s1","s2"),
                                       proportions = GRCh37.proportions, 
                                       file = file.path(tempdir(), "test.pdf"))
}
