% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID115_functions.R
\name{ReadID115Catalog}
\alias{ReadID115Catalog}
\title{Read catalog}
\usage{
ReadID115Catalog(
  file,
  ref.genome = NULL,
  region = "unknown",
  catalog.type = "counts"
)
}
\arguments{
\item{file}{Path to a catalog on disk in the standardized format.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{region A character string designating a genomic region;
see \code{\link{as.catalog}} and \code{\link{ICAMS}}.}

\item{catalog.type}{One of "counts", "density", "counts.signature",
"density.signature".}
}
\value{
A catalog as an S3 object; see \code{\link{as.catalog}}.
}
\description{
Read a catalog in standardized format from path.
}
\details{
See also \code{\link{WriteCatalog}}
}
\note{
In ID (small insertion and deletion) catalogs, deletion repeat sizes
range from 0 to 5+, but for plotting and end-user documentation
deletion repeat sizes range from 1 to 6+.
}
\section{Comments}{

To add or change attributes of the catalog, you can use function
\code{\link[base]{attr}}. \cr For example, \code{attr(catalog, "abundance")
<- custom.abundance}.
}

