% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Reverse}
\alias{Reverse}
\alias{reverse}
\title{Transcription bias of indels classified into 115 categories (purine)}
\format{
A \code{vector} which contains the 115 categories of indel events, but in purine
format

An object of class \code{character} of length 36.
}
\usage{
reverse
}
\description{
This data is designed to be used as an example in function \cr
\code{\link{PlotTransBiasID115}} and \code{\link{PlotTransBiasID115ToPdf}}.
}
\keyword{datasets}
