% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Reverse_pooled}
\alias{Reverse_pooled}
\alias{reverse_pooled}
\title{Transcription bias of indels classified into 13 categories (purine)}
\format{
A \code{vector} which contains the 13 categories of indel events, standardised
to purine format.

An object of class \code{character} of length 4.
}
\usage{
reverse_pooled
}
\description{
This data is designed to be used as an example in function \cr
\code{\link{ID115_PlotTransBias}} and
\code{\link{ID115_PlotTransBias}} when \code{pool = TRUE}.
}
\keyword{datasets}
