% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Target}
\alias{Target}
\alias{target}
\title{Transcription bias of indels classified into 115 categories (pyrimidine)}
\format{
A \code{vector} which contains the 115 categories of indel events, standardised
to pyrimidine format.

An object of class \code{character} of length 36.
}
\usage{
target
}
\description{
This data is designed to be used as an example in function \cr
\code{\link{ID115_PlotTransBias}} and \code{\link{ID115_PlotTransBiasToPdf}}.
}
\keyword{datasets}
