% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Target_pooled}
\alias{Target_pooled}
\alias{target_pooled}
\title{Transcription bias of indels classified into 13 categories (pyrimidine)}
\format{
A \code{vector} which contains the 13 categories of indel events, standardised
to pyrimidine format.

An object of class \code{character} of length 4.
}
\usage{
target_pooled
}
\description{
This data is designed to be used as an example in function \cr
\code{\link{PlotTransBiasID115}} and
\code{\link{PlotTransBiasID115}}
}
\keyword{datasets}
