% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.sigs.R
\name{TP_FP_FN_avg_sim}
\alias{TP_FP_FN_avg_sim}
\title{Return #TP, #FP, #FN, average cosine similarity between extracted
and ground truth signatures.}
\usage{
TP_FP_FN_avg_sim(extracted.sigs, ground.truth.sigs, similarity.cutoff = 0.9)
}
\arguments{
\item{extracted.sigs}{Mutational signatures discovered by some analysis.
A numerical-matrix-like object with columns as signatures.}

\item{ground.truth.sigs}{Ground-truth mutational signatures from
a synthetic data set. A numerical-matrix-like object with columns
as signatures.}

\item{similarity.cutoff}{A signature in \code{ground.truth.sigs}
must be matched
by \code{>= similarity.cutoff} by a signature in \code{extracted.sigs}
to be considered detected.}
}
\description{
Return #TP, #FP, #FN, average cosine similarity between extracted
and ground truth signatures.
}
\details{
Match signatures in \code{extracted.sigs} to
signatures in \code{ground.truth.sigs} using the function
\code{\link[clue]{solve_LSAP}}, which used the
"Hungarian" (a.k.a "Kuhn–Munkres") algorithm.
\url{https://en.wikipedia.org/wiki/Hungarian_algorithm},
which optimizes the total cost associated with the links
between nodes.
The function first computes the
all-pairs cosine similarity matrix between the two
sets of signatures, then converts cosine similarities
to cosine distances (including \code{similarity.cutoff})
by subtracting from 1, then
sets distances > the converted cutoff to very large values.
The applies \code{\link[clue]{solve_LSAP}} to the resulting
matrix to compute an optimal matching between
\code{extracted.sigs} and \code{ground.truth.sigs}.
}
