% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID115_functions.R
\name{VCFsToID115CatalogsAndPlotToPdf}
\alias{VCFsToID115CatalogsAndPlotToPdf}
\title{Read a list of vcfs and plot ID115 catalogs as pdf}
\usage{
VCFsToID115CatalogsAndPlotToPdf(
  list.of.vcfs,
  ref.genome,
  region = "unknown",
  flag.mismatches = 0,
  file,
  ylim = NULL
)
}
\arguments{
\item{list.of.vcfs}{List of in-memory ID VCFs. The list names will be
the sample ids in the output catalog.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}

\item{flag.mismatches}{Optional. If > 0, then if there are mismatches to
references in the ID (insertion/deletion) VCF, generate messages showing
the mismatched rows and continue. Otherwise \code{stop} if there are
mismatched rows. See \code{\link{AnnotateIDVCF}} for more details.}

\item{file}{The name of the PDF file to be produced.}

\item{ylim}{Has the usual meaning. Only implemented for SBS96Catalog and
IndelCatalog.}
}
\description{
Read a list of vcfs and plot ID115 catalogs as pdf
}
