% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID115_fns.R
\name{WriteID115Catalog}
\alias{WriteID115Catalog}
\title{Write a catalog to a file.}
\usage{
WriteID115Catalog(catalog, file, strict = TRUE)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{file}{The path of the file to be written.}

\item{strict}{If TRUE, then stop if additional checks on the input fail.}
}
\description{
Write a catalog to a file.
}
