% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.sigs.R
\name{MatchSigsAndRelabel}
\alias{MatchSigsAndRelabel}
\title{An asymmetrical analysis of a set of "ground truth" and "extracted" signatures.}
\usage{
MatchSigsAndRelabel(ex.sigs, gt.sigs, exposure = NULL, similarity.cutoff = 0.9)
}
\arguments{
\item{ex.sigs}{Newly extracted signatures to be compared to \code{gt.sigs}.}

\item{gt.sigs}{"Ground truth" signatures.}

\item{exposure}{If \code{NULL}, then match
\code{ex.sigs} against all signatures in \code{gt.sigs}.
Otherwise this should be ground-truth exposures used generate the
synthetic spectra from which \code{ex.sigs} were extracted.
In this case we do not
match to ground-truth signatures that were not in the ground
truth exposure.}

\item{similarity.cutoff}{A ground-truth signature must have been
the best match of an extracted signature with a cosine
similarity $ge$ this value to be considered a true positive.
Otherwise we conider the ground-truth signature to be a false
negative.}
}
\value{
A list with the elements
\itemize{
\item \code{averCosSim}
\item \code{match1}
\item \code{match2}
\item \code{extracted.with.no.best.match}
\item \code{ground.truth.with.no.best.match}
\item \code{ex.sigs}
\item \code{gt.sigs}
\item \code{gt.mean.cos.sim}

as for \code{\link{MatchSigs2Directions}},
with \code{match1} being matches for the the extracted signatures
(\code{ex.sigs}) and \code{match2} being the matches for
the ground truth signatures (\code{gt.sigs}). The return list
also echos the input arguments \code{ex.sigs} and \code{gt.sigs}.
}
}
\description{
An asymmetrical analysis of a set of "ground truth" and "extracted" signatures.
}
\examples{
gt.sigs <- matrix(c(1, 3,   4, 1, 2, 4), ncol = 2)
ex.sigs <- matrix(c(1, 3.1, 4, 5, 1, 1, 1, 2.8, 4), ncol = 3)
colnames(gt.sigs) <- c("gt.1", "gt.2")
colnames(ex.sigs) <- c("ex.1", "ex.2", "ex.3")
tout <- MatchSigsAndRelabel(gt.sigs = gt.sigs, ex.sigs = ex.sigs)
tout

}
